Use Long for representing nEvents
Because Int is not enough to specify the number of events.

+++ src/main/kotlin/org/fluentd/benchmark/BenchmarkConfig.kt
@@ -7,7 +7,7 @@ import java.nio.ByteBuffer
 
 class BenchmarkConfig(val tag: String,
                       val timestampType: BenchmarkClient.TimestampType,
-                      val nEvents: Int,
+                      val nEvents: Long,
                       val interval: Long?,
                       val period: Long?,
                       val recordKey: String,
@@ -60,7 +60,7 @@ class BenchmarkConfig(val tag: String,
 
         lateinit var tag: String
         var timestampType = BenchmarkClient.TimestampType.EventTime
-        var nEvents: Int = 10000
+        var nEvents: Long = 10000L
         var interval: Long? = null
         var period: Long? = null
         lateinit var recordKey: String
@@ -74,7 +74,7 @@ class BenchmarkConfig(val tag: String,
 
         fun timestampType(init: Builder.() -> BenchmarkClient.TimestampType) = apply { timestampType = init() }
 
-        fun nEvents(init: Builder.() -> Int) = apply { nEvents = init() }
+        fun nEvents(init: Builder.() -> Long) = apply { nEvents = init() }
 
         fun interval(init: Builder.() -> Long?) = apply { interval = init() }
